+++ pkg/scalers/metrics_api_scaler_test.go
@@ -10,6 +10,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
+
+	"github.com/kedacore/keda/v2/pkg/scalers/scalersconfig"
 )
 
 type metricsAPIMetadataTestData struct {
@@ -77,7 +79,7 @@ var testMetricsAPIAuthMetadata = []metricAPIAuthMetadataTestData{
 
 func TestParseMetricsAPIMetadata(t *testing.T) {
 	for _, testData := range testMetricsAPIMetadata {
-		_, err := parseMetricsAPIMetadata(&ScalerConfig{TriggerMetadata: testData.metadata, AuthParams: map[string]string{}})
+		_, err := parseMetricsAPIMetadata(&scalersconfig.ScalerConfig{TriggerMetadata: testData.metadata, AuthParams: map[string]string{}})
 		if err != nil && !testData.raisesError {
 			t.Error("Expected success but got error", err)
 		}
@@ -100,7 +102,7 @@ var metricsAPIMetricIdentifiers = []metricsAPIMetricIdentifier{
 func TestMetricsAPIGetMetricSpecForScaling(t *testing.T) {
 	for _, testData := range metricsAPIMetricIdentifiers {
 		s, err := NewMetricsAPIScaler(
-			&ScalerConfig{
+			&scalersconfig.ScalerConfig{
 				ResolvedEnv:       map[string]string{},
 				TriggerMetadata:   testData.metadataTestData.metadata,
 				AuthParams:        map[string]string{},
@@ -162,7 +164,7 @@ func TestGetValueFromResponse(t *testing.T) {
 
 func TestMetricAPIScalerAuthParams(t *testing.T) {
 	for _, testData := range testMetricsAPIAuthMetadata {
-		meta, err := parseMetricsAPIMetadata(&ScalerConfig{TriggerMetadata: testData.metadata, AuthParams: testData.authParams})
+		meta, err := parseMetricsAPIMetadata(&scalersconfig.ScalerConfig{TriggerMetadata: testData.metadata, AuthParams: testData.authParams})
 
 		if err != nil && !testData.isError {
 			t.Error("Expected success but got error", err)
@@ -208,7 +210,7 @@ func TestBearerAuth(t *testing.T) {
 	}
 
 	s, err := NewMetricsAPIScaler(
-		&ScalerConfig{
+		&scalersconfig.ScalerConfig{
 			ResolvedEnv:       map[string]string{},
 			TriggerMetadata:   metadata,
 			AuthParams:        authentication,